some refactor & minor updates

+++ app/Http/Controllers/LocationController.php
@@ -5,41 +5,31 @@ namespace App\Http\Controllers;
 use App\Http\Requests\LocationStoreRequest;
 use App\Http\Requests\LocationUpdateRequest;
 use App\Models\Location;
+use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 use Inertia\Inertia;
+use Inertia\Response;
 
 class LocationController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     */
-    public function index()
-    {
-        //
-    }
-
-    /**
-     * Show the form for creating a new resource.
-     */
-    public function create()
+    public function create(): Response
     {
         return Inertia::render('Location/Create');
     }
 
-    public function validate(Request $request)
+    public function validate(Request $request): void
     {
         $orig = new LocationStoreRequest();
 
         $request->validate($orig->rules(), $orig->messages());
     }
 
-    /**
-     * Store a newly created resource in storage.
-     */
-    public function store(LocationStoreRequest $request)
+    public function store(LocationStoreRequest $request): RedirectResponse
     {
         $validated = $request->validated();
+
+        $validated['schedule'] = implode(';', $validated['schedule']);
 
         $pathes = [];
         foreach ($validated['photos'] as $photo) {
@@ -48,33 +38,28 @@ class LocationController extends Controller
         }
         $validated['photos'] = implode(';', $pathes);
 
-        $request->user()->saloon->locations()->create($validated);
+        $request->user()->userable->locations()->create($validated);
+
+        return redirect(route('profile'));
     }
 
-    /**
-     * Display the specified resource.
-     */
     public function show(Location $location)
     {
         //
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     */
-    public function edit(Location $location)
+    public function edit(Location $location): Response
     {
         return Inertia::render('Location/Edit', [
             'location' => $location
         ]);
     }
 
-    /**
-     * Update the specified resource in storage.
-     */
-    public function update(LocationUpdateRequest $request, Location $location)
+    public function update(LocationUpdateRequest $request, Location $location): RedirectResponse
     {
         $validated = $request->validated();
+
+        $validated['schedule'] = implode(';', $validated['schedule']);
 
         $pathes = [];
         foreach ($validated['photos'] as $photo) {
@@ -100,13 +85,16 @@ class LocationController extends Controller
         
         $location->fill($validated);
         $location->save();
+
+        return redirect(route('profile'));
     }
 
-    /**
-     * Remove the specified resource from storage.
-     */
-    public function destroy(Location $location)
+    public function destroy(Request $request, Location $location): void
     {
+        if ($request->user()->cannot('delete', $location)) {
+            abort(403);
+        }
+
         $location->delete();
     }
 }
